Remove flow
I'm trying to simplify this project and remove dependencies to make it 
easier to work with. While flow provides a benefit, it adds a lot of 
complexity to the component. I'd rather get confidence the code is 
working by adding more tests later (which is something I'm more familiar 
with).

This was done with the `flow-remove-types` tool using the `--pretty` 
option and then running 'eslint --fix'.